Improved upgrade script.

+++ content/assets/scripts/upgrade-to-2.9.1.sh
@@ -1,6 +1,7 @@
 #! /usr/bin/env bash
 
 RED='\033[0;31m'
+GREEN='\033[0;32m'
 NC='\033[0m' # No Color
 
 # Everything must be executed with the normal user.
@@ -52,23 +53,32 @@ if [[ $(vercomp $PHP_VERSION 7.2.0 ; echo $?) -eq 2  ]]; then
 fi
 
 # Update the root repository of MONARC
+echo "${GREEN}Updating root of the project…${NC}"
 git pull origin master
+if [[ $? -ne 0 ]]; then
+    echo -e "${RED}Error when updating root of the project.${NC}"
+    exit 1
+fi
 
 # Clean old directories
+echo "${GREEN}Removing old vendor, module and cache directories…${NC}"
 rm -Rf vendor/
 rm -Rf module/
 rm -Rf data/cache
 
 # Create new required directories
+echo "${GREEN}Creating new module and cache directories…${NC}"
 mkdir -p module/Monarc
 mkdir -p data/cache
 mkdir -p data/DoctrineORMModule
 mkdir -p data/LazyServices/Proxy
 
-# Updated code MONARC PHP libraries and dependencies
+# Update PHP dependencies of MONARC
+echo "${GREEN}Installing new PHP backend…${NC}"
 composer install -o
 
 # Create new symlinks
+echo "${GREEN}Creating new symlinks…${NC}"
 cd module/Monarc
 ln -s ./../../vendor/monarc/core Core
 if [[ -d ./../../vendor/monarc/frontoffice ]]; then
@@ -79,6 +89,12 @@ fi
 cd ../..
 
 # Update all the project
+echo "${GREEN}Updating the whole project…${NC}"
 ./scripts/update-all.sh -c
+
+
+echo "${GREEN}Upgrade finished.${NC}"
+echo "Checks the permission of the folder data/"
+
 
 exit 0
